Reorganized classes to prepare for new geocoding
