+++ api/src/main/kotlin/fobo66/valiutchik/api/CurrencyRatesDataSource.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2022 Andrey Mukamolov
+ *    Copyright 2024 Andrey Mukamolov
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -15,6 +15,8 @@
  */
 
 package fobo66.valiutchik.api
+
+import fobo66.valiutchik.api.entity.Currency
 
 interface CurrencyRatesDataSource {
   suspend fun loadExchangeRates(city: String): Set<Currency>
