+++ api/src/main/kotlin/fobo66/valiutchik/api/CurrencyRatesDataSourceImpl.kt
@@ -16,6 +16,7 @@
 
 package fobo66.valiutchik.api
 
+import fobo66.valiutchik.api.entity.Currency
 import io.ktor.client.HttpClient
 import io.ktor.client.plugins.ResponseException
 import io.ktor.client.request.basicAuth
@@ -24,6 +25,8 @@ import io.ktor.client.statement.bodyAsChannel
 import io.ktor.http.path
 import io.ktor.utils.io.jvm.javaio.toInputStream
 import java.io.IOException
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.withContext
 
 const val BASE_URL = "https://admin.myfin.by/"
 
@@ -31,7 +34,8 @@ class CurrencyRatesDataSourceImpl(
   private val client: HttpClient,
   private val parser: CurrencyRatesParser,
   private val username: String,
-  private val password: String
+  private val password: String,
+  private val ioDispatcher: CoroutineDispatcher
 ) : CurrencyRatesDataSource {
 
   private val citiesMap: Map<String, String> by lazy {
@@ -45,10 +49,10 @@ class CurrencyRatesDataSourceImpl(
     )
   }
 
-  override suspend fun loadExchangeRates(city: String): Set<Currency> {
+  override suspend fun loadExchangeRates(city: String): Set<Currency> = withContext(ioDispatcher) {
     val cityIndex = citiesMap[city] ?: "1"
 
-    return try {
+    try {
       val response = client.get(BASE_URL) {
         url {
           path("outer", "authXml", cityIndex)
