+++ api/src/main/kotlin/fobo66/valiutchik/api/CurrencyRatesParser.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Andrey Mukamolov
+ *    Copyright 2024 Andrey Mukamolov
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -16,6 +16,7 @@
 
 package fobo66.valiutchik.api
 
+import fobo66.valiutchik.api.entity.Currency
 import java.io.InputStream
 
 /**
