+++ api/src/main/kotlin/fobo66/valiutchik/api/di/Dispatcher.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Andrey Mukamolov
+ *    Copyright 2024 Andrey Mukamolov
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -14,21 +14,8 @@
  *    limitations under the License.
  */
 
-package fobo66.valiutchik.api
+package fobo66.valiutchik.api.di
 
-/**
- * Currency model from XML
- */
-data class Currency(
-  val bankname: String = "",
-  val usdBuy: String = "",
-  val usdSell: String = "",
-  val eurBuy: String = "",
-  val eurSell: String = "",
-  val rurBuy: String = "",
-  val rurSell: String = "",
-  val plnBuy: String = "",
-  val plnSell: String = "",
-  val uahBuy: String = "",
-  val uahSell: String = ""
-)
+enum class Dispatcher {
+  IO
+}
