+++ api/src/main/kotlin/fobo66/valiutchik/api/entity/GeocodingResult.kt
@@ -14,8 +14,14 @@
  *    limitations under the License.
  */
 
-package fobo66.valiutchik.core.di
+package fobo66.valiutchik.api.entity
 
-internal enum class Dispatcher {
-  IO
-}
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class GeocodingResult(
+  @SerialName("features")
+  val features: List<Feature> = listOf()
+)
