+++ api/src/test/kotlin/fobo66/valiutchik/api/ApiKoinModuleTest.kt
@@ -18,6 +18,7 @@ package fobo66.valiutchik.api
 
 import fobo66.valiutchik.api.di.apiModule
 import fobo66.valiutchik.api.di.credentialsModule
+import fobo66.valiutchik.api.di.dispatchersModule
 import fobo66.valiutchik.api.di.networkModule
 import io.ktor.client.HttpClientConfig
 import io.ktor.client.engine.HttpClientEngine
@@ -26,6 +27,12 @@ import org.koin.core.annotation.KoinExperimentalAPI
 import org.koin.test.verify.verify
 
 class ApiKoinModuleTest {
+
+  @OptIn(KoinExperimentalAPI::class)
+  @Test
+  fun `check dispatchers module`() {
+    dispatchersModule.verify()
+  }
 
   @OptIn(KoinExperimentalAPI::class)
   @Test
