+++ data/src/main/kotlin/fobo66/valiutchik/core/di/DataModule.kt
@@ -16,6 +16,7 @@
 
 package fobo66.valiutchik.core.di
 
+import fobo66.valiutchik.api.di.Dispatcher
 import fobo66.valiutchik.api.di.apiModule
 import fobo66.valiutchik.core.model.datasource.AssetsDataSource
 import fobo66.valiutchik.core.model.datasource.AssetsDataSourceImpl
@@ -52,18 +53,13 @@ import fobo66.valiutchik.core.model.repository.MapRepositoryImpl
 import fobo66.valiutchik.core.model.repository.PreferenceRepository
 import fobo66.valiutchik.core.model.repository.PreferenceRepositoryImpl
 import fobo66.valiutchik.core.util.BankNameNormalizer
-import kotlinx.coroutines.Dispatchers
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.module.dsl.singleOf
 import org.koin.core.qualifier.qualifier
 import org.koin.dsl.module
 
-val coroutineDispatchersModule = module {
-  single(qualifier(Dispatcher.IO)) { Dispatchers.IO }
-}
-
 val dataSourcesModule = module {
-  includes(apiModule, systemModule, coroutineDispatchersModule, thirdPartyModule)
+  includes(apiModule, systemModule, thirdPartyModule)
 
   single<AssetsDataSource> {
     AssetsDataSourceImpl(get())
@@ -116,7 +112,7 @@ val repositoriesModule = module {
   }
 
   single<CurrencyRateRepository> {
-    CurrencyRateRepositoryImpl(get(), get(), get(), get(), get(qualifier(Dispatcher.IO)))
+    CurrencyRateRepositoryImpl(get(), get(), get(), get())
   }
 
   single<CurrencyRatesTimestampRepository> {
