+++ data/src/main/kotlin/fobo66/valiutchik/core/model/datasource/BestCourseDataSource.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2022 Andrey Mukamolov
+ *    Copyright 2024 Andrey Mukamolov
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 package fobo66.valiutchik.core.model.datasource
 
-import fobo66.valiutchik.api.Currency
+import fobo66.valiutchik.api.entity.Currency
 
 /**
  * Datasource for determining best exchange rates in the dataset
