+++ data/src/test/kotlin/fobo66/valiutchik/core/DataKoinModuleTest.kt
@@ -16,7 +16,6 @@
 
 package fobo66.valiutchik.core
 
-import fobo66.valiutchik.core.di.coroutineDispatchersModule
 import fobo66.valiutchik.core.di.dataSourcesModule
 import fobo66.valiutchik.core.di.repositoriesModule
 import fobo66.valiutchik.core.di.secretsModule
@@ -29,12 +28,6 @@ import org.koin.core.annotation.KoinExperimentalAPI
 import org.koin.test.verify.verify
 
 class DataKoinModuleTest {
-
-  @OptIn(KoinExperimentalAPI::class)
-  @Test
-  fun `check dispatchers module`() {
-    coroutineDispatchersModule.verify()
-  }
 
   @OptIn(KoinExperimentalAPI::class)
   @Test
