+++ data/src/test/kotlin/fobo66/valiutchik/core/fake/FakeBestCourseDataSource.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2022 Andrey Mukamolov
+ *    Copyright 2024 Andrey Mukamolov
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 package fobo66.valiutchik.core.fake
 
-import fobo66.valiutchik.api.Currency
+import fobo66.valiutchik.api.entity.Currency
 import fobo66.valiutchik.core.model.datasource.BestCourseDataSource
 
 class FakeBestCourseDataSource : BestCourseDataSource {
