+++ data/src/test/kotlin/fobo66/valiutchik/core/model/repository/CurrencyRateRepositoryTest.kt
@@ -43,8 +43,7 @@ class CurrencyRateRepositoryTest {
     bestCourseDataSource,
     persistenceDataSource,
     currencyRatesDataSource,
-    BankNameNormalizer(),
-    ioDispatcher
+    BankNameNormalizer()
   )
 
   @AfterEach
